Initial commit. Functions here create a group of all possible trees.
Trees create a group by defining a unit, an inverse and addition. The
functions in this module define all those for nodes, and then use the
node definitions to define them for trees. 

The result is an interesting module that allows you to, thanks to
group theory, perform a sort of tree arithmetic - add and subtract
trees and there is a "zero" tree (the unit). These functions should be
applicable to any xml tree.