Neighbor generation works, worked around bug; '-' in the tag algebra
We have to exclude tags with '-' in them because that character is not
included in the character domain of the Element.Tag module, which
these functions are based on. 
Adding '-'to the character domain is not an option; lxml does not
allow elements to be named just '-'. However, if that character is in
the domain then the group theory that these modules are based on
demands that elements can be named '-'.
The only way to solve this problem is to somehow force lxml to allow
elements to be named '-', but that is quite difficult. This was not
done. 
For now, elements whose tags contain '-' are simply excluded from
the acceptableTags list generated by the Errors module, and so such
elements are excluded from neighbor generation.