wrote tree ordering function using sax. Works, but useless.
The sax parser reads from disk, but trees are parsed into memory and
manipulated there. Using the sax function fails in compound operations
because changes are written to memory and not to disk, so the sax
ordering function does not see them. Fixing this would require redoing
a large part of the tree library.